refactor: consolidate column titles for pool and market list
