+++ apps/main/src/dex/features/pool-list/columns/index.ts
@@ -1,4 +1,5 @@
 export * from './columns.enum'
+export * from './column.titles'
 export * from './column.definitions'
 export * from './column.options'
 export * from './legacy-columns.enum'
