+++ apps/main/src/llamalend/features/market-list/columns/index.ts
@@ -1,4 +1,5 @@
 export * from './columns.enum'
+export * from './column.titles'
 export * from './column.options'
 export * from './column.constants'
 export * from './column.definitions'
